Refactor and enhance yasmin packages
- Removed unnecessary destructors from PclToRosPointCloud2State and RosToPclPointCloud2State classes.
- Added cancellation checks in SavePcdState and SavePlyState execute methods.
- Improved ActionState to notify when action is done and handle cancellation more effectively.
- Updated basic_outcomes to use inline constexpr for string constants.
- Enhanced MonitorState and ServiceState to include cancellation checks and improved locking mechanisms.
- Refactored ROSClientsCache to use per-map locks for thread safety.
- Updated get_parameters_state to handle int64_t instead of int for parameter declaration.
- Improved YasminViewerNode to manage active HTTP connections and handle server shutdown more gracefully.
- Refactored yasmin_viewer_pub to use unordered_map for better performance and added checks for rclpy state before publishing.

+++ yasmin_ros/src/yasmin_ros/get_parameters_state.cpp
@@ -74,7 +74,7 @@ GetParametersState::execute(yasmin::Blackboard::SharedPtr blackboard) {
       } else if (default_value.type() == typeid(std::vector<bool>)) {
         this->node_->declare_parameter(
             param_name, std::any_cast<std::vector<bool>>(default_value));
-      } else if (default_value.type() == typeid(std::vector<int>)) {
+      } else if (default_value.type() == typeid(std::vector<int64_t>)) {
         this->node_->declare_parameter(
             param_name, std::any_cast<std::vector<int64_t>>(default_value));
       } else if (default_value.type() == typeid(std::vector<double>)) {
